[review] xtest: case TEEC_CONFIG_SHAREDMEM_MAX_SIZE == ULONG_MAX
Size shall be size_t, not uint32_t.

+++ host/xtest/regression_5000.c
@@ -78,7 +78,7 @@ struct xtest_session {
 
 /* Registers the TEEC_SharedMemory to the TEE. */
 static TEEC_Result RegisterSharedMemory(TEEC_Context *ctx,
-					TEEC_SharedMemory *shm, uint32_t size,
+					TEEC_SharedMemory *shm, size_t size,
 					uint32_t flags)
 {
 	shm->flags = flags;
@@ -88,7 +88,7 @@ static TEEC_Result RegisterSharedMemory(TEEC_Context *ctx,
 
 /* Allocates shared memory inside of the TEE. */
 static TEEC_Result AllocateSharedMemory(TEEC_Context *ctx,
-					TEEC_SharedMemory *shm, uint32_t size,
+					TEEC_SharedMemory *shm, size_t size,
 					uint32_t flags)
 {
 	shm->flags = flags;
@@ -111,7 +111,7 @@ static void Allocate_In(struct xtest_session *cs)
 	Do_ADBG_BeginSubCase(cs->c, "Allocate_In");
 	{
 		TEEC_SharedMemory shm;
-		uint32_t size = 1024;
+		size_t size = 1024;
 
 		if (!ADBG_EXPECT(cs->c, TEEC_SUCCESS,
 			TEEC_InitializeContext(_device, &cs->context)))
@@ -135,7 +135,7 @@ static void Allocate_out_of_memory(struct xtest_session *cs)
 	Do_ADBG_BeginSubCase(cs->c, "Allocate_out_of_memory");
 	{
 		TEEC_SharedMemory shm;
-		uint32_t SIZE_OVER_MEMORY_CAPACITY = INT32_MAX;
+		size_t SIZE_OVER_MEMORY_CAPACITY = INT32_MAX;
 
 		if (!ADBG_EXPECT(cs->c, TEEC_SUCCESS,
 			TEEC_InitializeContext(_device, &cs->context)))
@@ -352,7 +352,7 @@ static void AllocateThenRegister_SameMemory(struct xtest_session *cs)
 	Do_ADBG_BeginSubCase(cs->c, "AllocateThenRegister_SameMemory");
 	{
 		TEEC_SharedMemory shm;
-		uint32_t size_allocation = 32;
+		size_t size_allocation = 32;
 
 		if (!ADBG_EXPECT(cs->c, TEEC_SUCCESS,
 			TEEC_InitializeContext(_device, &cs->context)))
@@ -380,7 +380,7 @@ static void AllocateSameMemory_twice(struct xtest_session *cs)
 	Do_ADBG_BeginSubCase(cs->c, "AllocateSameMemory_twice");
 	{
 		TEEC_SharedMemory shm;
-		uint32_t size_allocation = 32;
+		size_t size_allocation = 32;
 
 		if (!ADBG_EXPECT(cs->c, TEEC_SUCCESS,
 			TEEC_InitializeContext(_device, &cs->context)))
@@ -434,7 +434,7 @@ out:
 
 static void Allocate_sharedMemory_maxSize(struct xtest_session *cs)
 {
-	uint32_t size_max = MIN(TEE_CLIENT_API_LOWER_SHM_BUFFER_MAX_SIZE,
+	size_t size_max = MIN(TEE_CLIENT_API_LOWER_SHM_BUFFER_MAX_SIZE,
 				TEEC_CONFIG_SHAREDMEM_MAX_SIZE);
 
 	Do_ADBG_BeginSubCase(cs->c,
@@ -462,7 +462,7 @@ out:
 
 static void Allocate_sharedMemory_belowMaxSize(struct xtest_session *cs)
 {
-	uint32_t size_below = MIN(TEE_CLIENT_API_LOWER_SHM_BUFFER_MAX_SIZE,
+	size_t size_below = MIN(TEE_CLIENT_API_LOWER_SHM_BUFFER_MAX_SIZE,
 				  TEEC_CONFIG_SHAREDMEM_MAX_SIZE) - 1;
 
 	Do_ADBG_BeginSubCase(cs->c,
@@ -492,7 +492,7 @@ out:
 
 static void Allocate_sharedMemory_aboveMaxSize(struct xtest_session *cs)
 {
-	uint32_t size_above = MIN(TEE_CLIENT_API_LOWER_SHM_BUFFER_MAX_SIZE,
+	size_t size_above = MIN(TEE_CLIENT_API_LOWER_SHM_BUFFER_MAX_SIZE,
 				  TEEC_CONFIG_SHAREDMEM_MAX_SIZE) + 1;
 
 	Do_ADBG_BeginSubCase(cs->c,
@@ -524,7 +524,7 @@ out:
 
 static void Register_sharedMemory_maxSize(struct xtest_session *cs)
 {
-	uint32_t size_max = MIN(TEE_CLIENT_API_LOWER_SHM_BUFFER_MAX_SIZE,
+	size_t size_max = MIN(TEE_CLIENT_API_LOWER_SHM_BUFFER_MAX_SIZE,
 				TEEC_CONFIG_SHAREDMEM_MAX_SIZE);
 
 	Do_ADBG_BeginSubCase(cs->c, "Register_sharedMemory_maxSize");
@@ -555,7 +555,7 @@ static void Register_sharedMemory_aboveMaxSize(struct xtest_session *cs)
 	Do_ADBG_BeginSubCase(cs->c, "Register_sharedMemory_aboveMaxSize");
 	{
 		TEEC_Result res;
-		uint32_t size_aboveMax = UINT32_MAX;
+		size_t size_aboveMax = UINT32_MAX;
 		uint8_t val[1];
 		TEEC_SharedMemory shm;
 
@@ -580,7 +580,7 @@ out:
 
 static void Register_sharedMemory_belowMaxSize(struct xtest_session *cs)
 {
-	uint32_t size_belowMax = MIN(TEEC_CONFIG_SHAREDMEM_MAX_SIZE,
+	size_t size_belowMax = MIN(TEEC_CONFIG_SHAREDMEM_MAX_SIZE,
 			     TEE_CLIENT_API_LOWER_SHM_BUFFER_MAX_SIZE) - 1;
 
 	Do_ADBG_BeginSubCase(cs->c, "Register_sharedMemory_belowMaxSize");
